refactor(all): extract virtual repeater strategy interface
